extract binary-tree/treenode pkg
